Generic array to dictionary function and improved fromList.
Improved JSON parser styling. Allow make to be public for encoders. Allow resolveKeypath to be public for resolving keypaths on JSONValue.

Refactored to use mapKeyValues as an extension to array. Updated JSON make/resolvedKeypaths to be private again.

Cleaner Dictionary from list of Key-Value of pairs using initializer on Dictionary. Added mapLabelPairs and mapLabel to convert a SequenceType or array to a dictionary of Key-Value pairs.